requirements.txt
fastapi
uvicorn
fpdf2>=2.7.6